Created a rough framework for the app navigation.
A UITableViewController is the main navigation page, and has a UIImageView as a header.
There are two UITableViewControllers for exhibit and event navigation, and currently just have a label for a name.
A UIViewController will hold more information, currently has a single text view.

No segues have been created between these yet.